Decode Parameter Bugfix
The decode function did not have a params argument, and when called with paramters would always error out claiming that the required parameters were not submitted.

Fixed by adding the params argument to decode.

Note: The recieve subcommand does not pass a third argument to decode. It seems it should - comments?